Respond to API review feedback
Remove abstraction for ResourceReader/Writer: just reuse the source.

Remove non-essential members.
