--- pkg/test/packageSettings/System.Resources.Extensions/settings.targets
@@ -1,6 +1,0 @@
-<Project>
-  <ItemGroup>
-    <IgnoredTypes Include="System.Resources.ResourceReader" />
-    <IgnoredTypes Include="System.Resources.ResourceWriter" />
-  </ItemGroup>
-</Project>
