+++ src/Common/src/CoreLib/System/Resources/FastResourceComparer.cs
@@ -15,7 +15,6 @@
 ** 
 ===========================================================*/
 
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
