+++ src/Common/src/CoreLib/System/Resources/ResourceReader.cs
@@ -17,21 +17,20 @@
 ===========================================================*/
 
 namespace System.Resources
+#if RESOURCES_STANDALONE
+    .Extensions
+#endif
 {
     using System;
     using System.IO;
     using System.Text;
     using System.Collections;
     using System.Collections.Generic;
-    using System.Reflection;
-    using System.Security;
-    using System.Globalization;
-    using System.Configuration.Assemblies;
-    using System.Runtime.Versioning;
     using System.Diagnostics;
-    using System.Diagnostics.Contracts;
-    using System.Threading;
 
+#if RESOURCES_STANDALONE
+    using ResourceReader = DeserializingResourceReader;
+#endif
     // Provides the default implementation of IResourceReader, reading
     // .resources file from the system default binary format.  This class
     // can be treated as an enumerator once.
@@ -72,7 +71,13 @@ namespace System.Resources
         }
     }
 
-    public partial class ResourceReader : IResourceReader
+    public sealed partial class
+#if RESOURCES_STANDALONE
+        DeserializingResourceReader
+#else
+        ResourceReader
+#endif
+        : IResourceReader
     {
         // A reasonable default buffer size for reading from files, especially
         // when we will likely be seeking frequently.  Could be smaller, but does
@@ -100,15 +105,6 @@ namespace System.Resources
         private int[] _typeNamePositions = null!;  // To delay initialize type table
         private int _numResources;    // Num of resources files, in case arrays aren't allocated.
 
-        private readonly bool _permitDeserialization;  // can deserialize BinaryFormatted resources
-        private object? _binaryFormatter; // binary formatter instance to use for deserializing
-
-        // statics used to dynamically call into BinaryFormatter
-        // When successfully located s_binaryFormatterType will point to the BinaryFormatter type
-        // and s_deserializeMethod will point to an unbound delegate to the deserialize method.
-        private static Type s_binaryFormatterType;
-        private static Func<object?, Stream, object> s_deserializeMethod;
-
         // We'll include a separate code path that uses UnmanagedMemoryStream to
         // avoid allocating String objects and the like.
         private UnmanagedMemoryStream? _ums;
@@ -117,7 +113,13 @@ namespace System.Resources
         private int _version;
 
 
-        public ResourceReader(string fileName)
+        public
+
+#if RESOURCES_STANDALONE
+        DeserializingResourceReader(string fileName)
+#else
+        ResourceReader(string fileName)
+#endif
         {
             _resCache = new Dictionary<string, ResourceLocator>(FastResourceComparer.Default);
             _store = new BinaryReader(new FileStream(fileName, FileMode.Open, FileAccess.Read, FileShare.Read, DefaultFileStreamBufferSize, FileOptions.RandomAccess), Encoding.UTF8);
@@ -133,7 +135,13 @@ namespace System.Resources
             }
         }
 
-        public ResourceReader(Stream stream)
+        public
+#if RESOURCES_STANDALONE
+        DeserializingResourceReader(Stream stream)
+#else
+        ResourceReader(Stream stream)
+#endif
+
         {
             if (stream == null)
                 throw new ArgumentNullException(nameof(stream));
@@ -147,27 +155,6 @@ namespace System.Resources
 
             ReadResources();
         }
-
-        // This is the constructor the RuntimeResourceSet calls,
-        // passing in the stream to read from and the RuntimeResourceSet's 
-        // internal hash table (hash table of names with file offsets
-        // and values, coupled to this ResourceReader).
-        internal ResourceReader(Stream stream, Dictionary<string, ResourceLocator> resCache, bool permitDeserialization)
-        {
-            Debug.Assert(stream != null, "Need a stream!");
-            Debug.Assert(stream.CanRead, "Stream should be readable!");
-            Debug.Assert(resCache != null, "Need a Dictionary!");
-
-            _resCache = resCache;
-            _store = new BinaryReader(stream, Encoding.UTF8);
-
-            _ums = stream as UnmanagedMemoryStream;
-
-            _permitDeserialization = permitDeserialization;
-
-            ReadResources();
-        }
-
 
         public void Close()
         {
@@ -202,7 +189,6 @@ namespace System.Resources
             }
         }
 
-        protected virtual int Version { get => 2; } // File format version number
 
         internal static unsafe int ReadUnalignedI4(int* p)
         {
@@ -772,71 +758,6 @@ namespace System.Resources
             return DeserializeObject(typeIndex);
         }
 
-        private object DeserializeObject(int typeIndex)
-        {
-            if (!_permitDeserialization && GetType() == typeof(ResourceReader))
-            {
-                throw new NotSupportedException(SR.NotSupported_ResourceObjectSerialization);
-            }
-
-            Type type = FindType(typeIndex);
-
-            object graph = DeserializeObject(_store, type);
-            
-            // guard against corrupted resources
-            if (graph.GetType() != type)
-                throw new BadImageFormatException(SR.Format(SR.BadImageFormat_ResType_SerBlobMismatch, type.FullName, graph.GetType().FullName));
- 
-            return graph;
-        }
-
-        protected virtual object DeserializeObject(BinaryReader reader, Type type)
-        {
-            if (!_permitDeserialization)
-                throw new NotSupportedException(SR.NotSupported_ResourceObjectSerialization);
-            if (reader == null)
-                throw new ArgumentNullException(nameof(reader));
-            if (type == null)
-                throw new ArgumentNullException(nameof(type));
-
-            if (_binaryFormatter == null)
-            {
-                InitializeBinaryFormatter();
-            }
-
-            return s_deserializeMethod(_binaryFormatter, reader.BaseStream);
-        }
-
-        private void InitializeBinaryFormatter()
-        {
-            LazyInitializer.EnsureInitialized(ref s_binaryFormatterType, () =>
-                Type.GetType("System.Runtime.Serialization.Formatters.Binary.BinaryFormatter, System.Runtime.Serialization.Formatters, Version=0.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a",
-                throwOnError: true));
-
-            LazyInitializer.EnsureInitialized(ref s_deserializeMethod, () =>
-               {
-                   MethodInfo binaryFormatterDeserialize = s_binaryFormatterType.GetMethod("Deserialize", new Type[] { typeof(Stream) });
-
-                    // create an unbound delegate that can accept a BinaryFormatter instance as object
-                    return (Func<object?, Stream, object>)typeof(ResourceReader)
-                            .GetMethod(nameof(CreateUntypedDelegate), BindingFlags.NonPublic | BindingFlags.Static)
-                            .MakeGenericMethod(s_binaryFormatterType)
-                            .Invoke(null, new object[] { binaryFormatterDeserialize });
-               });
-
-            _binaryFormatter = Activator.CreateInstance(s_binaryFormatterType);
-        }
-
-        // generic method that we specialize at runtime once we've loaded the BinaryFormatter type
-        // permits creating an unbound delegate so that we can avoid reflection after the initial
-        // lightup code completes.
-        private static Func<object, Stream, object> CreateUntypedDelegate<TInstance>(MethodInfo method)
-        {
-            Func<TInstance, Stream, object> typedDelegate = (Func<TInstance, Stream, object>)Delegate.CreateDelegate(typeof(Func<TInstance, Stream, object>), null, method);
-
-            return (obj, stream) => typedDelegate((TInstance)obj, stream);
-        }
-
         // Reads in the header information for a .resources file.  Verifies some
         // of the assumptions about this resource set, and builds the class table
         // for the default resource file format.
@@ -890,7 +811,7 @@ namespace System.Resources
                 string readerType = _store.ReadString();
                 string thisType = GetType().AssemblyQualifiedName;
 
-                if (!CompareNames(readerType, thisType))
+                if (!ValidateReaderType(readerType))
                     throw new NotSupportedException(SR.Format(SR.NotSupported_WrongResourceReader_Type, readerType));
 
                 // Skip over type name for a suitable ResourceSet
@@ -1059,72 +980,6 @@ namespace System.Resources
             return _typeTable[typeIndex]!; // TODO-NULLABLE: https://github.com/dotnet/roslyn/issues/34644
         }
 
-        public void GetResourceData(string resourceName, out string resourceType, out byte[] resourceData)
-        {
-            if (resourceName == null)
-                throw new ArgumentNullException(nameof(resourceName));
-            if (_resCache == null)
-                throw new InvalidOperationException(SR.ResourceReaderIsClosed);
-
-            // Get the type information from the data section.  Also,
-            // sort all of the data section's indexes to compute length of
-            // the serialized data for this type (making sure to subtract
-            // off the length of the type code).
-            int[] sortedDataPositions = new int[_numResources];
-            int dataPos = FindPosForResource(resourceName);
-            if (dataPos == -1)
-            {
-                throw new ArgumentException(SR.Format(SR.Arg_ResourceNameNotExist, resourceName));
-            }
-
-            lock (this)
-            {
-                // Read all the positions of data within the data section.
-                for (int i = 0; i < _numResources; i++)
-                {
-                    _store.BaseStream.Position = _nameSectionOffset + GetNamePosition(i);
-                    // Skip over name of resource
-                    int numBytesToSkip = _store.Read7BitEncodedInt();
-                    if (numBytesToSkip < 0)
-                    {
-                        throw new FormatException(SR.Format(SR.BadImageFormat_ResourcesNameInvalidOffset, numBytesToSkip));
-                    }
-                    _store.BaseStream.Position += numBytesToSkip;
-
-                    int dPos = _store.ReadInt32();
-                    if (dPos < 0 || dPos >= _store.BaseStream.Length - _dataSectionOffset)
-                    {
-                        throw new FormatException(SR.Format(SR.BadImageFormat_ResourcesDataInvalidOffset, dPos));
-                    }
-                    sortedDataPositions[i] = dPos;
-                }
-                Array.Sort(sortedDataPositions);
-
-                int index = Array.BinarySearch(sortedDataPositions, dataPos);
-                Debug.Assert(index >= 0 && index < _numResources, "Couldn't find data position within sorted data positions array!");
-                long nextData = (index < _numResources - 1) ? sortedDataPositions[index + 1] + _dataSectionOffset : _store.BaseStream.Length;
-                int len = (int)(nextData - (dataPos + _dataSectionOffset));
-                Debug.Assert(len >= 0 && len <= (int)_store.BaseStream.Length - dataPos + _dataSectionOffset, "Length was negative or outside the bounds of the file!");
-
-                // Read type code then byte[]
-                _store.BaseStream.Position = _dataSectionOffset + dataPos;
-                ResourceTypeCode typeCode = (ResourceTypeCode)_store.Read7BitEncodedInt();
-                if (typeCode < 0 || typeCode >= ResourceTypeCode.StartOfUserTypes + _typeTable.Length)
-                {
-                    throw new BadImageFormatException(SR.BadImageFormat_InvalidType);
-                }
-                resourceType = TypeNameFromTypeCode(typeCode);
-
-                // The length must be adjusted to subtract off the number 
-                // of bytes in the 7 bit encoded type code.
-                len -= (int)(_store.BaseStream.Position - (_dataSectionOffset + dataPos));
-                byte[] bytes = _store.ReadBytes(len);
-                if (bytes.Length != len)
-                    throw new FormatException(SR.BadImageFormat_ResourceNameCorrupted);
-                resourceData = bytes;
-            }
-        }
-
         private string TypeNameFromTypeCode(ResourceTypeCode typeCode)
         {
             Debug.Assert(typeCode >= 0, "can't be negative");
@@ -1148,47 +1003,6 @@ namespace System.Resources
                     _store.BaseStream.Position = oldPos;
                 }
             }
-        }
-
-        private static bool CompareNames(string typeName1, string typeName2)
-        {
-            // First, compare type names
-            int comma1 = typeName1.IndexOf(',');
-            int comma2 = typeName2.IndexOf(',');
-            if (comma1 != comma2)
-                return false;
-
-            // both are missing assembly name, compare entire string as type name
-            if (comma1 == -1)
-                return string.Equals(typeName1, typeName2, StringComparison.Ordinal);
-
-            // compare the type name portion
-            ReadOnlySpan<char> type1 = typeName1.AsSpan(0, comma1);
-            ReadOnlySpan<char> type2 = typeName2.AsSpan(0, comma2);
-            if (!type1.Equals(type2, StringComparison.Ordinal))
-                return false;
-
-            // Now, compare assembly display names (IGNORES VERSION AND PROCESSORARCHITECTURE)
-            // also, for  mscorlib ignores everything, since that's what the binder is going to do
-            while (Char.IsWhiteSpace(typeName1[++comma1]));
-            while (Char.IsWhiteSpace(typeName2[++comma2]));
-
-            // case insensitive
-            AssemblyName an1 = new AssemblyName(typeName1.Substring(comma1));
-            AssemblyName an2 = new AssemblyName(typeName1.Substring(comma2));
-            if (!string.Equals(an1.Name, an2.Name, StringComparison.OrdinalIgnoreCase))
-                return false;
-
-            // to match IsMscorlib() in VM
-            if (string.Equals(an1.Name, "mscorlib", StringComparison.OrdinalIgnoreCase))
-                return true;
-
-            if (an1.CultureInfo?.LCID != an2.CultureInfo?.LCID)
-                return false;
-
-            byte[] pkt1 = an1.GetPublicKeyToken();
-            byte[] pkt2 = an2.GetPublicKeyToken();
-            return pkt1.AsSpan().SequenceEqual(pkt2);
         }
 
         internal sealed class ResourceEnumerator : IDictionaryEnumerator
