+++ src/System.Resources.Extensions/System.Resources.Extensions.sln
@@ -16,64 +16,22 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{E5D875A9
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		netcoreapp-Debug|Any CPU = netcoreapp-Debug|Any CPU
-		netcoreapp-Release|Any CPU = netcoreapp-Release|Any CPU
 		netstandard-Debug|Any CPU = netstandard-Debug|Any CPU
 		netstandard-Release|Any CPU = netstandard-Release|Any CPU
-		Release|Any CPU = Release|Any CPU
-		uap-Debug|Any CPU = uap-Debug|Any CPU
-		uap-Release|Any CPU = uap-Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.Debug|Any CPU.ActiveCfg = netstandard-Debug|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.Debug|Any CPU.Build.0 = netstandard-Debug|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.netcoreapp-Debug|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.netcoreapp-Debug|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.netcoreapp-Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.netcoreapp-Release|Any CPU.Build.0 = netstandard-Release|Any CPU
 		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.netstandard-Debug|Any CPU.ActiveCfg = netstandard-Debug|Any CPU
 		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.netstandard-Debug|Any CPU.Build.0 = netstandard-Debug|Any CPU
 		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.netstandard-Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
 		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.netstandard-Release|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.Release|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.uap-Debug|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.uap-Debug|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.uap-Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{694389E4-6F68-4E0A-A8AC-2B85416C0742}.uap-Release|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.Debug|Any CPU.ActiveCfg = uap-Debug|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.Debug|Any CPU.Build.0 = uap-Debug|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.netcoreapp-Debug|Any CPU.ActiveCfg = netcoreapp-Debug|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.netcoreapp-Debug|Any CPU.Build.0 = netcoreapp-Debug|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.netcoreapp-Release|Any CPU.ActiveCfg = netcoreapp-Release|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.netcoreapp-Release|Any CPU.Build.0 = netcoreapp-Release|Any CPU
 		{CD6088BF-DDA2-4097-A505-D786E8968A76}.netstandard-Debug|Any CPU.ActiveCfg = netstandard-Debug|Any CPU
 		{CD6088BF-DDA2-4097-A505-D786E8968A76}.netstandard-Debug|Any CPU.Build.0 = netstandard-Debug|Any CPU
 		{CD6088BF-DDA2-4097-A505-D786E8968A76}.netstandard-Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
 		{CD6088BF-DDA2-4097-A505-D786E8968A76}.netstandard-Release|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.Release|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.uap-Debug|Any CPU.ActiveCfg = uap-Debug|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.uap-Debug|Any CPU.Build.0 = uap-Debug|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.uap-Release|Any CPU.ActiveCfg = uap-Release|Any CPU
-		{CD6088BF-DDA2-4097-A505-D786E8968A76}.uap-Release|Any CPU.Build.0 = uap-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.Debug|Any CPU.ActiveCfg = netstandard-Debug|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.Debug|Any CPU.Build.0 = netstandard-Debug|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.netcoreapp-Debug|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.netcoreapp-Debug|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.netcoreapp-Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.netcoreapp-Release|Any CPU.Build.0 = netstandard-Release|Any CPU
 		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.netstandard-Debug|Any CPU.ActiveCfg = netstandard-Debug|Any CPU
 		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.netstandard-Debug|Any CPU.Build.0 = netstandard-Debug|Any CPU
 		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.netstandard-Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
 		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.netstandard-Release|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.Release|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.uap-Debug|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.uap-Debug|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.uap-Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
-		{9359E7DF-50C8-471B-AF9E-706FE2C4B326}.uap-Release|Any CPU.Build.0 = netstandard-Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
