+++ src/System.Resources.Extensions/ref/Configurations.props
@@ -2,8 +2,6 @@
   <PropertyGroup>
     <BuildConfigurations>
       netstandard;
-      netcoreapp;
-      uap;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
