+++ src/System.Resources.Extensions/src/BinaryReaderExtensions.cs
@@ -26,20 +26,4 @@
             return count;
         }
     }
-
-    internal static class BinaryWriterExtensions
-    {
-        public static void Write7BitEncodedInt(this BinaryWriter writer, int value)
-        {
-            // Write out an int 7 bits at a time.  The high bit of the byte,
-            // when on, tells reader to continue reading more bytes.
-            uint v = (uint)value;   // support negative numbers
-            while (v >= 0x80)
-            {
-                writer.Write((byte)(v | 0x80));
-                v >>= 7;
-            }
-            writer.Write((byte)v);
-        }
-    }
 }
