+++ src/System.Resources.Extensions/src/System/Resources/Extensions/SerializationFormat.cs
@@ -1,3 +1,6 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
 namespace System.Resources.Extensions
 {
@@ -6,6 +9,6 @@ namespace System.Resources.Extensions
         BinaryFormatter = 0x1,
         TypeConverterByteArray = 0x2,
         TypeConverterString = 0x3,
-        Stream = 0x4
+        ActivatorStream = 0x4
     }
 }
